+++ packages/evmos-wallet/src/wallet/ButtonWalletConnection.tsx
@@ -164,6 +164,7 @@ export const ButtonWalletConnection = ({
             <button
               className="w-full rounded font-bold uppercase border border-darkPearl hover:bg-grayOpacity p-3 mt-3"
               onClick={() => {
+                RemoveProviderFromLocalStorage();
                 disconnectWallets(dispatch);
                 setShow(false);
                 setIsCopied(false);